+++ YarnSpinner.Compiler/TypeCheckerListener.cs
@@ -8,8 +8,18 @@ namespace Yarn.Compiler
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Threading;
     using TypeChecker;
     using Yarn.Utility;
+
+    internal abstract class InterruptableListener : YarnSpinnerParserBaseListener
+    {
+        protected CancellationToken CancellationToken { get; private set; }
+        protected InterruptableListener(CancellationToken cancellationToken)
+        {
+            this.CancellationToken = cancellationToken;
+        }
+    }
 
     /// <summary>
     /// <see cref="TypeCheckerListener"/> creates type constraints (subclasses
@@ -19,7 +29,7 @@ namespace Yarn.Compiler
     /// the types of expressions and values in the source code (or to determine
     /// if there was a type error.)
     /// </summary>
-    internal class TypeCheckerListener : YarnSpinnerParserBaseListener
+    internal class TypeCheckerListener : InterruptableListener
     {
         /// <summary>
         /// Stores the total number of type variables that have been produced.
@@ -135,7 +145,7 @@ namespace Yarn.Compiler
         /// cref="TypeConstraint"/> objects. During type checking, this
         /// collection will be added to if a type constraint fails to resolve by
         /// the end of a file.</param>
-        public TypeCheckerListener(string sourceFileName, CommonTokenStream tokens, List<Declaration> knownDeclarations, List<TypeBase> knownTypes, Substitution typeSolution, HashSet<TypeConstraint> failingTypeConstraints)
+        public TypeCheckerListener(string sourceFileName, CommonTokenStream tokens, List<Declaration> knownDeclarations, List<TypeBase> knownTypes, Substitution typeSolution, HashSet<TypeConstraint> failingTypeConstraints, CancellationToken cancellationToken) : base(cancellationToken)
         {
             this.sourceFileName = sourceFileName;
             this.tokens = tokens;
@@ -1202,6 +1212,8 @@ namespace Yarn.Compiler
 
         public override void ExitStatement([NotNull] YarnSpinnerParser.StatementContext context)
         {
+            this.CancellationToken.ThrowIfCancellationRequested();
+
             try
             {
                 if (this.typeEquations.Count == 0)